Return response as stream of bytes
In order to handle non-text output in the message body, the whole response
should be returned as a stream of bytes. socket/give function changed to
BufferedOutputStream rather than BufferedReader. Function in response namespace
added to handle byte conversions, and response/compose amended to turn the
response message and body into lazy sequence of bytes. operator/content method
similarly written to get file content as bytes.

Also significant reduction of operator unit tests to reflect that the hard-coded
routes are now in a hash-map in router namespace. Kept the tests that actually
reflected the branching logic in the operator functions.