Add logging functionality, turn on dotall flag for regex
Found that unit tests were passing but acceptance tests were not; culprit lay
with failure to turn on dotall flag for regex on parsing requests.
Unit tests forgot to cover
multiple headers condition, which is why this situation was not caught there.
Upside is that inputs simulating requests have been checked for consistency
across all unit tests and with HTTP specification.

In the process of attempting to diagnose the problem, logging all requests was
implemented in addition to a clear-logs function, although the latter is not
used yet.