Parse request body, implement partial content
Request message body is returned as part of hash-map from request/parse. Route
handler for requests on /form written so that POST and PUT message body is
written to a file called form, DELETE removes the file, and GET always returns
200 even if the file does not exist.

Router can handle GET requests with Range header asking for partial content. The
header is parsed to determine the start and end indices for the content. The
response/content function rewritten to have optional parameters for start and
end indices.

The httpserver.authentication namespace renamed to encoding. File handling
boolean functions like not-found? and directory? moved to their own namespace
in httpserver.file.