Extend router-fn as parameter to serve and threading functions
Default behavior will be to assume there is no additional custom routes
specified. But arg can be supplied to -main via -r flag and parameter provided
when calling threading and serve functions. Note that this parameter needs to be
explicitly set as nil when calling threading and serve directly without wanting
to provide a function for custom routes.